+++ crates/integrations/connector-integration/src/payout_connectors.rs
@@ -1,0 +1,8 @@
+pub mod itaubank;
+pub use self::itaubank::ItaubankPayouts;
+
+pub mod loonio;
+pub use self::loonio::LoonioPayouts;
+
+pub mod paypal;
+pub use self::paypal::PaypalPayouts;