+++ crates/types-traits/domain_types/src/connector_flow.rs
@@ -97,6 +97,12 @@ pub struct PayoutEnrollDisburseAccount;
 #[derive(Debug, Clone)]
 pub struct SurchargeCalculate;
 
+#[derive(Debug, Clone)]
+pub struct SurchargePaymentSucceeded;
+
+#[derive(Debug, Clone)]
+pub struct SurchargeRefundSucceeded;
+
 #[derive(strum::Display)]
 #[strum(serialize_all = "snake_case")]
 pub enum FlowName {
@@ -134,4 +140,6 @@ pub enum FlowName {
     PayoutCreateRecipient,
     PayoutEnrollDisburseAccount,
     SurchargeCalculate,
+    SurchargePaymentSucceeded,
+    SurchargeRefundSucceeded,
 }